Fix Ruby on AWS Lambda

+++ cli/src/providers/aws_lambda.rs
@@ -1,5 +1,6 @@
 use std::fs;
 use std::io::Read;
+use std::path::Path;
 use std::rc::Rc;
 use std::sync::Arc;
 
@@ -41,7 +42,9 @@ impl AwsLambdaProvider {
     pub fn cast_iomods(ctx: Rc<Context>, service_name: &str) -> Result<(), CastError> {
         let project_path = ctx.project.path.clone();
         let iomod_path = format!("{}/net/services/{}/iomods", project_path, service_name);
-        fs::remove_dir_all(iomod_path.clone()).expect("could not rm iomod directory");
+        if Path::new(&iomod_path.clone()).exists() {
+            fs::remove_dir_all(iomod_path.clone()).expect("could not rm iomod directory");
+        }
         fs::create_dir_all(iomod_path.clone()).expect("could not create iomod directory");
 
         let mut dependencies: Vec<String> = Vec::new();
@@ -83,12 +86,21 @@ impl AwsLambdaProvider {
 
         archive::zip_files(
             dependencies,
-            format!("./.asml/runtime/{}.zip", &service_name),
+            format!("./.asml/runtime/{}-iomods.zip", &service_name),
             Some("iomod/"),
             false,
         );
 
         Ok(())
+    }
+
+    pub fn cast_ruby(ctx: Rc<Context>, service_name: &str) -> Result<(), CastError> {
+        let project_path = ctx.project.path.clone();
+        let ruby_dir = format!("{}/net/services/{}/ruby-wasm32-wasi", project_path, service_name);
+        archive::zip_dir(
+            ruby_dir.into(),
+            format!("./.asml/runtime/{}-ruby.zip", &service_name)
+        ).map_err(|_| CastError("could not zip ruby env directory".into()))
     }
 }
 
@@ -166,9 +178,14 @@ impl Castable for LambdaService {
         );
 
         AwsLambdaProvider::cast_iomods(ctx.clone(), &name).unwrap();
+        let mut has_ruby_layer = false;
+        if ctx.functions.iter().find(|f| f.language == "ruby").is_some() {
+            AwsLambdaProvider::cast_ruby(ctx.clone(), &name)?;
+            has_ruby_layer = true;
+        }
 
         let use_apigw = ctx.functions.iter().find(|f| f.http.is_some()).is_some();
-        let has_service_layer = ctx.iomods.len() > 0;
+        let has_iomods_layer = ctx.iomods.len() > 0;
 
         let authorizers: Vec<ServiceAuthData> = ctx
             .authorizers
@@ -196,7 +213,8 @@ impl Castable for LambdaService {
             service_name: name.clone(),
             layer_name,
             use_apigw,
-            has_service_layer,
+            has_iomods_layer,
+            has_ruby_layer,
             authorizers,
             options: self.options.clone(),
         }.render();
@@ -297,12 +315,19 @@ impl Castable for LambdaFunction {
                         "aws_lambda_layer_version.asml_{}_runtime.arn",
                         service.clone()
                     ),
-                    service_layer: match iomod_names.len() {
+                    iomods_layer: match iomod_names.len() {
                         0 => None,
                         _ => Some(format!(
-                            "aws_lambda_layer_version.asml_{}_service.arn",
+                            "aws_lambda_layer_version.asml_{}_iomods.arn",
                             service.clone()
                         )),
+                    },
+                    ruby_layer: match &*function.language {
+                        "ruby" => Some(format!(
+                            "aws_lambda_layer_version.asml_{}_ruby.arn",
+                            service.clone()
+                        )),
+                        _ => None,
                     },
                     size: function.size,
                     timeout: function.timeout,
@@ -362,7 +387,8 @@ struct ServiceTemplate {
     project_name: String,
     service_name: String,
     layer_name: String,
-    has_service_layer: bool,
+    has_iomods_layer: bool,
+    has_ruby_layer: bool,
     use_apigw: bool,
     authorizers: Vec<ServiceAuthData>,
     options: Arc<Options>,
@@ -388,13 +414,22 @@ resource aws_lambda_layer_version asml_{{service_name}}_runtime {
     source_code_hash = filebase64sha256("${local.project_path}/.asml/runtime/bootstrap.zip")
 }
 
-{{#if has_service_layer}}resource aws_lambda_layer_version asml_{{service_name}}_service {
+{{#if has_iomods_layer}}resource aws_lambda_layer_version asml_{{service_name}}_iomods {
     provider = aws.{{project_name}}
 
-    filename   = "${local.project_path}/.asml/runtime/{{service_name}}.zip"
-    layer_name = "asml-${local.project_name}-{{service_name}}-service"
+    filename   = "${local.project_path}/.asml/runtime/{{service_name}}-iomods.zip"
+    layer_name = "asml-${local.project_name}-{{service_name}}-iomods"
 
-    source_code_hash = filebase64sha256("${local.project_path}/.asml/runtime/{{service_name}}.zip")
+    source_code_hash = filebase64sha256("${local.project_path}/.asml/runtime/{{service_name}}-iomods.zip")
+}{{/if}}
+
+{{#if has_ruby_layer}}resource aws_lambda_layer_version asml_{{service_name}}_ruby {
+    provider = aws.{{project_name}}
+
+    filename   = "${local.project_path}/.asml/runtime/{{service_name}}-ruby.zip"
+    layer_name = "asml-${local.project_name}-{{service_name}}-ruby"
+
+    source_code_hash = filebase64sha256("${local.project_path}/.asml/runtime/{{service_name}}-ruby.zip")
 }{{/if}}
 
 {{#if use_apigw}}resource aws_apigatewayv2_api {{service_name}}_http_api {
@@ -433,7 +468,8 @@ pub struct FunctionTemplate {
     pub service_name: String,
     pub function_name: String,
     pub runtime_layer: String,
-    pub service_layer: Option<String>,
+    pub iomods_layer: Option<String>,
+    pub ruby_layer: Option<String>,
     pub http: Option<HttpData>,
     pub auth: Option<FunctionAuthData>,
     pub size: u16,
@@ -463,7 +499,13 @@ resource aws_lambda_function asml_{{service_name}}_{{function_name}} {
     timeout       = {{timeout}}
     memory_size   = {{size}}
 
-    layers = [{{runtime_layer}}{{#if service_layer}}, {{service_layer}}{{/if}}]
+    {{#if ruby_layer}}environment {
+      variables = {
+        ASML_FUNCTION_ENV = "ruby-lambda"
+      }
+    }{{/if}}
+
+    layers = [{{runtime_layer}}{{#if iomods_layer}}, {{iomods_layer}}{{/if}}{{#if ruby_layer}}, {{ruby_layer}}{{/if}}]
 
     source_code_hash = filebase64sha256("${local.project_path}/net/services/{{service_name}}/{{function_name}}/{{function_name}}.zip")
 }
